[INLONG-9866][Sort] Workflow test, no need merge.
