+++ inlong-sort/sort-end-to-end-tests/sort-end-to-end-tests-v1.15/src/test/java/org/apache/inlong/sort/tests/Mongodb2StarRocksTesta.java
@@ -60,9 +60,9 @@ import static com.mongodb.client.model.Updates.*;
  * End-to-end tests for sort-connector-mongodb-cdc-v1.15 uber jar.
  * Test flink sql Mongodb cdc to StarRocks
  */
-public class Mongodb2StarRocksTest extends FlinkContainerTestEnv {
+public class Mongodb2StarRocksTesta extends FlinkContainerTestEnv {
 
-    private static final Logger LOG = LoggerFactory.getLogger(Mongodb2StarRocksTest.class);
+    private static final Logger LOG = LoggerFactory.getLogger(Mongodb2StarRocksTesta.class);
 
     private static final Path mongodbJar = TestUtils.getResource("sort-connector-mongodb-cdc.jar");
     private static final Path jdbcJar = TestUtils.getResource("sort-connector-starrocks.jar");
@@ -82,7 +82,7 @@ public class Mongodb2StarRocksTest extends FlinkContainerTestEnv {
 
     static {
         try {
-            sqlFile = Paths.get(Postgres2StarRocksTest.class.getResource("/flinkSql/mongodb_test.sql").toURI())
+            sqlFile = Paths.get(Postgres2StarRocksTesta.class.getResource("/flinkSql/mongodb_test.sql").toURI())
                     .toString();
             buildStarRocksImage();
         } catch (URISyntaxException e) {
