+++ inlong-sort/sort-end-to-end-tests/sort-end-to-end-tests-v1.15/src/test/java/org/apache/inlong/sort/tests/Mysql2StarRocksTesta.java
@@ -52,9 +52,9 @@ import static org.apache.inlong.sort.tests.utils.StarRocksManager.initializeStar
  * End-to-end tests for sort-connector-postgres-cdc-v1.15 uber jar.
  * Test flink sql Mysql cdc to StarRocks
  */
-public class Mysql2StarRocksTest extends FlinkContainerTestEnv {
+public class Mysql2StarRocksTesta extends FlinkContainerTestEnv {
 
-    private static final Logger LOG = LoggerFactory.getLogger(Mysql2StarRocksTest.class);
+    private static final Logger LOG = LoggerFactory.getLogger(Mysql2StarRocksTesta.class);
 
     private static final Path mysqlJar = TestUtils.getResource("sort-connector-mysql-cdc.jar");
     private static final Path jdbcJar = TestUtils.getResource("sort-connector-starrocks.jar");
@@ -64,7 +64,7 @@ public class Mysql2StarRocksTest extends FlinkContainerTestEnv {
     static {
         try {
             sqlFile =
-                    Paths.get(Mysql2StarRocksTest.class.getResource("/flinkSql/mysql_test.sql").toURI()).toString();
+                    Paths.get(Mysql2StarRocksTesta.class.getResource("/flinkSql/mysql_test.sql").toURI()).toString();
             buildStarRocksImage();
         } catch (URISyntaxException e) {
             throw new RuntimeException(e);
