+++ inlong-sort/sort-end-to-end-tests/sort-end-to-end-tests-v1.15/src/test/java/org/apache/inlong/sort/tests/Postgres2StarRocksTesta.java
@@ -19,7 +19,6 @@ package org.apache.inlong.sort.tests;
 
 import org.apache.inlong.sort.tests.utils.FlinkContainerTestEnv;
 import org.apache.inlong.sort.tests.utils.JdbcProxy;
-import org.apache.inlong.sort.tests.utils.MSSQLServerContainer;
 import org.apache.inlong.sort.tests.utils.StarRocksContainer;
 import org.apache.inlong.sort.tests.utils.TestUtils;
 
@@ -29,6 +28,7 @@ import org.junit.ClassRule;
 import org.junit.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.testcontainers.containers.PostgreSQLContainer;
 import org.testcontainers.containers.output.Slf4jLogConsumer;
 import org.testcontainers.utility.DockerImageName;
 
@@ -48,21 +48,22 @@ import static org.apache.inlong.sort.tests.utils.StarRocksManager.STAR_ROCKS_LOG
 import static org.apache.inlong.sort.tests.utils.StarRocksManager.buildStarRocksImage;
 import static org.apache.inlong.sort.tests.utils.StarRocksManager.getNewStarRocksImageName;
 import static org.apache.inlong.sort.tests.utils.StarRocksManager.initializeStarRocksTable;
+/**
+ * End-to-end tests for sort-connector-postgres-cdc-v1.15 uber jar.
+ * Test flink sql Postgres cdc to StarRocks
+ */
+public class Postgres2StarRocksTesta extends FlinkContainerTestEnv {
 
-public class Sqlserver2StarRocksTest extends FlinkContainerTestEnv {
+    private static final Logger LOG = LoggerFactory.getLogger(Postgres2StarRocksTesta.class);
 
-    private static final Logger LOG = LoggerFactory.getLogger(Sqlserver2StarRocksTest.class);
-
-    private static final Path sqlserverJar = TestUtils.getResource("sort-connector-sqlserver-cdc.jar");
+    private static final Path postgresJar = TestUtils.getResource("sort-connector-postgres-cdc.jar");
     private static final Path jdbcJar = TestUtils.getResource("sort-connector-starrocks.jar");
-
-    private static final Path mysqlJar = TestUtils.getResource("mysql-driver.jar");
-
+    private static final Path mysqlJdbcJar = TestUtils.getResource("mysql-driver.jar");
     private static final String sqlFile;
 
     static {
         try {
-            sqlFile = Paths.get(Sqlserver2StarRocksTest.class.getResource("/flinkSql/sqlserver_test.sql").toURI())
+            sqlFile = Paths.get(Postgres2StarRocksTesta.class.getResource("/flinkSql/postgres_test.sql").toURI())
                     .toString();
             buildStarRocksImage();
         } catch (URISyntaxException e) {
@@ -80,55 +81,38 @@ public class Sqlserver2StarRocksTest extends FlinkContainerTestEnv {
                     .withLogConsumer(new Slf4jLogConsumer(STAR_ROCKS_LOG));
 
     @ClassRule
-    public static final MSSQLServerContainer SQLSERVER_CONTAINER = (MSSQLServerContainer) new MSSQLServerContainer(
-            DockerImageName.parse("mcr.microsoft.com/mssql/server").withTag("2022-latest"))
-                    .acceptLicense()
+    public static final PostgreSQLContainer POSTGRES_CONTAINER = (PostgreSQLContainer) new PostgreSQLContainer(
+            DockerImageName.parse("debezium/postgres:13").asCompatibleSubstituteFor("postgres"))
+                    .withUsername("flinkuser")
+                    .withPassword("flinkpw")
+                    .withDatabaseName("test")
                     .withNetwork(NETWORK)
-                    .withNetworkAliases("sqlserver")
+                    .withNetworkAliases("postgres")
                     .withLogConsumer(new Slf4jLogConsumer(LOG));
 
     @Before
     public void setup() {
         waitUntilJobRunning(Duration.ofSeconds(30));
-        initializeSqlserverTable();
+        initializePostgresTable();
         initializeStarRocksTable(STAR_ROCKS);
     }
 
-    private void initializeSqlserverTable() {
+    private void initializePostgresTable() {
         try {
-            // Waiting for MSSQL Agent started.
-            LOG.info("Sleep until the MSSQL Agent is started...");
-            Thread.sleep(20 * 1000);
-            LOG.info("Now continue initialize task...");
-            Class.forName(SQLSERVER_CONTAINER.getDriverClassName());
+            Class.forName(POSTGRES_CONTAINER.getDriverClassName());
             Connection conn = DriverManager
-                    .getConnection(SQLSERVER_CONTAINER.getJdbcUrl(), SQLSERVER_CONTAINER.getUsername(),
-                            SQLSERVER_CONTAINER.getPassword());
+                    .getConnection(POSTGRES_CONTAINER.getJdbcUrl(), POSTGRES_CONTAINER.getUsername(),
+                            POSTGRES_CONTAINER.getPassword());
             Statement stat = conn.createStatement();
-            stat.execute("CREATE DATABASE test;");
-            stat.execute("USE test;");
             stat.execute(
-                    "CREATE TABLE test_input1 (\n" +
-                            "    id INT IDENTITY(1,1) NOT NULL PRIMARY KEY,\n" +
-                            "    name NVARCHAR(255) NOT NULL DEFAULT 'flink',\n" +
-                            "    description NVARCHAR(512)\n" +
-                            ");");
-            stat.execute("if exists(select 1 from sys.databases where name='test' and is_cdc_enabled=0)\n" +
-                    "begin\n" +
-                    "    exec sys.sp_cdc_enable_db\n" +
-                    "end");
-            stat.execute("IF EXISTS(SELECT 1 FROM sys.tables WHERE name='test_input1' AND is_tracked_by_cdc = 0)\n" +
-                    "BEGIN\n" +
-                    "    EXEC sys.sp_cdc_enable_table\n" +
-                    "        @source_schema = 'dbo', -- source_schema\n" +
-                    "        @source_name = 'test_input1', -- table_name\n" +
-                    "        @capture_instance = NULL, -- capture_instance\n" +
-                    "        @supports_net_changes = '1', -- capture_instance\n" +
-                    "        @index_name = NULL,  -- \n" +
-                    "        @captured_column_list  = NULL, -- \n" +
-                    "        @filegroup_name = 'PRIMARY', -- \n" +
-                    "        @role_name = NULL -- role_name\n" +
-                    "END");
+                    "CREATE TABLE test_input1 (\n"
+                            + "  id SERIAL,\n"
+                            + "  name VARCHAR(255) NOT NULL DEFAULT 'flink',\n"
+                            + "  description VARCHAR(512),\n"
+                            + "  PRIMARY  KEY(id)\n"
+                            + ");");
+            stat.execute(
+                    "ALTER TABLE test_input1 REPLICA IDENTITY FULL; ");
             stat.close();
             conn.close();
         } catch (Exception e) {
@@ -138,8 +122,8 @@ public class Sqlserver2StarRocksTest extends FlinkContainerTestEnv {
 
     @AfterClass
     public static void teardown() {
-        if (SQLSERVER_CONTAINER != null) {
-            SQLSERVER_CONTAINER.stop();
+        if (POSTGRES_CONTAINER != null) {
+            POSTGRES_CONTAINER.stop();
         }
         if (STAR_ROCKS != null) {
             STAR_ROCKS.stop();
@@ -152,26 +136,20 @@ public class Sqlserver2StarRocksTest extends FlinkContainerTestEnv {
      * @throws Exception The exception may throws when execute the case
      */
     @Test
-    public void testSqlserverUpdateAndDelete() throws Exception {
-        submitSQLJob(sqlFile, jdbcJar, sqlserverJar, mysqlJar);
+    public void testPostgresUpdateAndDelete() throws Exception {
+        submitSQLJob(sqlFile, jdbcJar, postgresJar, mysqlJdbcJar);
         waitUntilJobRunning(Duration.ofSeconds(10));
 
         // generate input
         try (Connection conn =
-                DriverManager.getConnection(SQLSERVER_CONTAINER.getJdbcUrl(), SQLSERVER_CONTAINER.getUsername(),
-                        SQLSERVER_CONTAINER.getPassword());
+                DriverManager.getConnection(POSTGRES_CONTAINER.getJdbcUrl(), POSTGRES_CONTAINER.getUsername(),
+                        POSTGRES_CONTAINER.getPassword());
                 Statement stat = conn.createStatement()) {
-            stat.execute("USE test;");
             stat.execute(
-                    "SET IDENTITY_INSERT test_input1 ON;" +
-                            "INSERT INTO test_input1 (id, name, description) "
-                            + "VALUES (1, 'jacket','water resistent white wind breaker');" +
-                            "SET IDENTITY_INSERT test_input1 OFF;");
+                    "INSERT INTO test_input1 "
+                            + "VALUES (1,'jacket','water resistent white wind breaker');");
             stat.execute(
-                    "SET IDENTITY_INSERT test_input1 ON;" +
-                            "INSERT INTO test_input1 (id, name, description) " +
-                            "VALUES (2,'scooter','Big 2-wheel scooter ');" +
-                            "SET IDENTITY_INSERT test_input1 OFF;");
+                    "INSERT INTO test_input1 VALUES (2,'scooter','Big 2-wheel scooter ');");
             stat.execute(
                     "update test_input1 set name = 'tom' where id = 2;");
             stat.execute(
